feat: move mpi to cpu backend + rename
