+++ src/backend/compute-backend.c
@@ -6,10 +6,6 @@
 #include "logger/log.h"
 #include <stdlib.h>
 
-#ifdef USE_MPI
-#include "mpi/mpi-backend.h"
-#endif
-
 struct compute_backend *compute_backend_create(struct p_args *args, struct filter_mix *filters)
 {
 	enum conv_backend type = args->compute_cfg.backend;
@@ -32,15 +28,6 @@ struct compute_backend *compute_backend_create(struct p_args *args, struct filte
 	case CONV_BACKEND_GPU:
 		backend->ops= &gpu_backend_ops;
 		break;
-	case CONV_BACKEND_MPI: 
-#ifdef USE_MPI
-		backend->ops = &mpi_backend_ops;
-		break;
-#else
-		log_error("Error: MPI backend requested but USE_MPI is not defined\n");
-		free(backend);
-		return NULL;
-#endif
 	default:
 		log_error("Error: Unknown compute backend type %d\n", type);
 		free(backend);
