+++ src/utils/threads-general.c
@@ -340,7 +340,7 @@ void save_result_image(char *output_filepath, size_t path_len, int threadnum, bm
 
 	if (strcmp(args->files_cfg.output_filename, "") != 0) {
 		snprintf(output_filepath, path_len, "test-img/%s", args->files_cfg.output_filename);
-	} else if (args->compute_cfg.backend == CONV_BACKEND_MPI) {
+	} else if (args->compute_cfg.mpi == CONV_MPI_ENABLED) {
 		snprintf(output_filepath, path_len, "test-img/mpi_out_%s", args->files_cfg.input_filename[0]);
 	} else {
 		if (threadnum > 1) {
